Move class to its own file
